ml/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
ultralytics>=8.0.0
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
opencv-python>=4.8.0
tensorrt>=8.6.0; platform_system == "Linux"
pycuda>=2022.1; platform_system == "Linux"
